+++ R/plot_continuous.R
@@ -1,7 +1,8 @@
 #' Plot weights across a continuous covariate
 #'
-#' This provides a plot showing the marginal distribution of a covariate
-#' under the implied regression weights and without weights.
+#' This provides a simple plot for the distribution of a single
+#' continuous covariate in the nominal sample and the implicit sample
+#' defined by the Aronow and Samii (2015) regression weights.
 #' @param mod Weighting model object
 #' @param covariate Covariate vector
 #' @param alpha Number between zero and one indicating the desired alpha level